Updated Bioreactor Variables and Calculations
